add commented-out attempt at installing Ruby 1.8.7
I cannot get it to work now, it won't be easy to get rbenv + Ruby 1.8 on
Fedora 17.

However, there seems to be a fix when using RVM:
http://shlrm.org/blog/2012/07/02/building-ruby-1-dot-8-7-on-fedora-17-using-rvm/

Unfortunately rbenv doesn't support patching in the process of Ruby
installation. I guess Ruby would have to be built with manual
interception. This might be useful:
http://stackoverflow.com/questions/6134456/error-while-installing-ruby-1-8-7-on-fedora-15